fix(lobbies): add protection for unexpected disconnects

+++ app/consumers.py
@@ -53,6 +53,7 @@ class LobbyConsumer(AsyncWebsocketConsumer):
     def init_player(self, player_name):
         player = self.lobby.players.filter(name=player_name).first()
         player.channel_name = self.channel_name
+        player.is_disconnected = False
         player.save()
         channel_to_lobby[self.channel_name] = player.lobby.name
 
@@ -63,6 +64,8 @@ class LobbyConsumer(AsyncWebsocketConsumer):
     # mark player as disconnected
     @database_sync_to_async
     def disconnect_player(self):
+        if not self.lobby:
+            return
         player = self.lobby.players.filter(channel_name=self.channel_name).first()
         if not player or player.is_disconnected:
             return
@@ -84,7 +87,8 @@ class LobbyConsumer(AsyncWebsocketConsumer):
 
     @database_sync_to_async
     def delete_lobby(self):
-       self.lobby.delete()
+       if self.lobby:
+           self.lobby.delete()
 
     async def send_players_list(self):
         players, max_players = await self.get_players()
@@ -101,18 +105,25 @@ class LobbyConsumer(AsyncWebsocketConsumer):
         await self.accept()
 
         # when a player connect to a lobby, send the list players to all connected players
-        await self.send_players_list()
+        try:
+            await self.send_players_list()
+        except:
+            pass
 
     async def disconnect(self, _):
         await self.disconnect_player()
         await self.channel_layer.group_discard(self.lobby_name, self.channel_name)
-        if await self.is_lobby_empty():
+        if not self.lobby or await self.is_lobby_empty():
            await self.delete_lobby()
            if self.task:
                self.task.cancel()
         else:
-            await self.send_players_list()
-        channel_to_lobby.pop(self.channel_name)
+            try:
+                await self.send_players_list()
+            except:
+                pass
+        if self.channel_name in channel_to_lobby.keys():
+            channel_to_lobby.pop(self.channel_name)
 
     async def receive(self, text_data):
         data = json.loads(text_data)
@@ -179,7 +190,7 @@ class LobbyConsumer(AsyncWebsocketConsumer):
         )
 
     def end_game(self, _):
-        if lobby_to_gs[self.lobby.name]:
+        if self.lobby.name in lobby_to_gs.keys():
            lobby_to_gs.pop(self.lobby.name)
 
     async def mark_ai(self, gs):
@@ -226,7 +237,10 @@ class LobbyConsumer(AsyncWebsocketConsumer):
             self.tournament = Tournament(gs, alive_player_names, gs.is_four_player)
             self.tournament.get_match_count()
             self.tournament.start_tournament()
-            self.task = asyncio.create_task(self.run_matches())
+            try:
+                self.task = asyncio.create_task(self.run_matches())
+            except:
+                pass
             self.task.add_done_callback(self.end_game)
 
     def get_game_state(self):
